Finished Minesweeper without clock
Finished Minesweeper, did not add how long it took the user to play the game. When player loses or wins, game pauses for 1000 ms then displays end screen and returns to main menu.
Extra space above the board,  reserved space for timer and mines left, could also change the color of the background and fix the bits of white shown on the numbers